planner-ai/pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "planner-ai"
version = "0.1.0"
description = "LangGraph-based travel planner AI service"
requires-python = ">=3.11"
dependencies = [
    "langgraph>=0.2.0",
    "langchain>=0.3.0",
    "langchain-core>=0.3.0",
    "langchain-openai>=0.3.0",
    "langchain-anthropic>=0.3.0",
    "langsmith>=0.2.0",
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.32.0",
    "asyncpg>=0.30.0",
    "redis[asyncio]>=5.2.0",
    "pydantic>=2.10.0",
    "python-dotenv>=1.0.0",
    "httpx>=0.28.0",
    "google-search-results>=2.4.2",
    "loguru>=0.7.0",
    "rich>=13.0.0",
    "pytest>=8.0.0",
    "pytest-asyncio>=0.24.0",
]


[tool.pytest.ini_options]
asyncio_mode = "auto"
